requirements.txt
elementpath
Django==3.2.*
google-api-core==2.8.2
google-api-python-client==2.51.0
google-auth==2.8.0
google-auth-httplib2==0.1.0
google-auth-oauthlib==0.5.2
googleapis-common-protos==1.56.3
lxml==4.9.0
oauth2client==4.1.3
oauthlib==3.2.0
psycopg2-binary==2.9.3
requests==2.28.0
schedule==1.1.0
urllib3==1.26.9